+++ packages/node-wot-servient-launcher/package.json
@@ -22,11 +22,12 @@
     "wot-servient": "dist/wos-servient.js"
   },
   "dependencies": {
-    "node-wot-protocols-http-server": "^0.1.0",
-    "node-wot-protocols-http-client": "^0.1.0",
-    "node-wot-protocols-coap-server": "^0.1.0",
-    "node-wot-protocols-coap-client": "^0.1.0",
     "node-wot-logger": "^0.1.0",
-    "node-wot-servient": "^0.1.0"
+    "node-wot-protocols-coap-client": "^0.1.0",
+    "node-wot-protocols-coap-server": "^0.1.0",
+    "node-wot-protocols-http-client": "^0.1.0",
+    "node-wot-protocols-http-server": "^0.1.0",
+    "node-wot-servient": "^0.1.0",
+    "wot-typescript-definitions": "^0.2.3"
   }
 }
